refactor(hub): move PendingFileRequests out of http into its own module
Round 2 S15: PendingFileRequests was defined in src/http/files.rs but
referenced by JobRuntime::handle_device_frame in the WS layer
(src/http/jobs.rs) and by AppState (src/state.rs). That coupled
WebSocket frame handling to an HTTP endpoint module, which is a
layering inversion — the HTTP and WS layers should both depend on a
transport-neutral piece of state.

Move PendingFileRequests + PendingFileRequestsError +
MAX_PENDING_FILE_REQUESTS + new_pending_requests + the inline
#[cfg(test)] mod into a new top-level module
`crate::pending_file_requests`. http/files.rs now imports
PendingFileRequestsError from there; state.rs and http/jobs.rs do the
same for PendingFileRequests itself.

No behavior change. All 6 inline unit tests still pass under the new
module path; the http_files integration tests are unaffected.
